appBackend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
dataclasses==0.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.2.31
dnspython==2.6.1
mongoengine==0.29.1
msgpack==1.1.0
PyJWT==2.9.0
pymongo==4.10.1
pytz==2024.2
redis==5.1.1
sqlparse==0.5.1
